@@ -1,10 +1,16 @@
-using Jitter.LinearMath;
 using Microsoft.Xna.Framework;
 
 namespace TestMonoGame.Extensions;
 
 public static class VectorExtensions
 {
+    public static void Set(this ref Vector3 vector, float x, float y, float z)
+    {
+        vector.X = x;
+        vector.Y = y;
+        vector.Z = z;
+    }
+    
     public static Vector3 ProjectOntoPlane(this Vector3 vector, Vector3 planeNormal)
     {
         // Ensure the plane normal is normalized
@@ -17,21 +23,4 @@ public static class VectorExtensions
         return projectedVector;
     }
     
-    public static JVector ToJVector(this Vector3 vector)
-    {
-        return new JVector(vector.X, vector.Y, vector.Z);
-    }
-
-    public static Vector3 ToVector3(this JVector vector)
-    {
-        return new Vector3(vector.X, vector.Y, vector.Z);
-    }
-
-    public static void FromJVector(this ref Vector3 vector, JVector jVector)
-    {
-        vector.X = jVector.X;
-        vector.Y = jVector.Y;
-        vector.Z = jVector.Z;
-    }
-
 }
